+++ swing/framework-ui/src/test/java/is/codion/swing/framework/ui/component/EntityComboBoxPanelTest.java
@@ -46,7 +46,8 @@ public final class EntityComboBoxPanelTest {
 	void test() {
 		EntityComboBoxModel model = EntityComboBoxModel.builder(Department.TYPE, CONNECTION_PROVIDER).build();
 		model.items().refresh();
-		ComponentValue<Entity, EntityComboBoxPanel> value = EntityComboBoxPanel.builder(model, () -> null)
+		ComponentValue<Entity, EntityComboBoxPanel> value = EntityComboBoxPanel.builder(model)
+						.editPanel(() -> null)
 						.buildValue();
 		Entity sales = CONNECTION_PROVIDER.connection().selectSingle(
 						Department.NAME.equalTo("SALES"));