+++ changelog.md
@@ -63,6 +63,7 @@ Codion Change Log
 - FilterTable.Builder.Model added, replacing the tableModel parameter in FilterTable.builder().
 ### is.codion.swing.framework.ui
 - ApplicationPreferences.fromString() bug fixed, FRAME_MAXIMIZED_KEY now used.
+- EntityComboBoxPanel.Builder.EditPanel added, replacing the editPanel parameter in EntityComboBoxPanel.builder().
 ### is.codion.framework.lambda
 - LambdaEntityHandler refactored.